+++ compose.yml
@@ -92,7 +92,6 @@ services:
       - node_modules_ghost_referrers:/home/ghost/ghost/referrers/node_modules:delegated
       - node_modules_ghost_security:/home/ghost/ghost/security/node_modules:delegated
       - node_modules_ghost_session-service:/home/ghost/ghost/session-service/node_modules:delegated
-      - node_modules_ghost_settings-path-manager:/home/ghost/ghost/settings-path-manager/node_modules:delegated
       - node_modules_ghost_slack-notifications:/home/ghost/ghost/slack-notifications/node_modules:delegated
       - node_modules_ghost_stripe:/home/ghost/ghost/stripe/node_modules:delegated
       - node_modules_ghost_tiers:/home/ghost/ghost/tiers/node_modules:delegated
@@ -251,7 +250,6 @@ volumes:
   node_modules_ghost_referrers: {}
   node_modules_ghost_security: {}
   node_modules_ghost_session-service: {}
-  node_modules_ghost_settings-path-manager: {}
   node_modules_ghost_slack-notifications: {}
   node_modules_ghost_stripe: {}
   node_modules_ghost_tiers: {}
