+++ ghost/core/core/server/services/route-settings/index.js
@@ -1,6 +1,5 @@
 const config = require('../../../shared/config');
 const parseYaml = require('./yaml-parser');
-const SettingsPathManager = require('@tryghost/settings-path-manager');
 
 let settingsLoader;
 let routeSettings;
@@ -10,6 +9,7 @@ module.exports = {
         const RouteSettings = require('./RouteSettings');
         const SettingsLoader = require('./SettingsLoader');
         const DefaultSettingsManager = require('./DefaultSettingsManager');
+        const SettingsPathManager = require('./SettingsPathManager');
 
         const settingsPathManager = new SettingsPathManager({type: 'routes', paths: [config.getContentPath('settings')]});
         settingsLoader = new SettingsLoader({parseYaml, settingFilePath: settingsPathManager.getDefaultFilePath()});
