+++ ghost/core/package.json
@@ -141,7 +141,6 @@
     "@tryghost/request": "1.0.8",
     "@tryghost/security": "0.0.0",
     "@tryghost/session-service": "0.0.0",
-    "@tryghost/settings-path-manager": "0.0.0",
     "@tryghost/slack-notifications": "0.0.0",
     "@tryghost/social-urls": "0.1.43",
     "@tryghost/string": "0.2.12",
@@ -166,6 +165,7 @@
     "connect-slashes": "1.4.0",
     "cookie-session": "2.1.0",
     "cors": "2.8.5",
+    "date-fns": "2.30.0",
     "dompurify": "3.2.3",
     "downsize": "0.0.8",
     "express": "4.21.2",
